+++ src/hooks/queries/useWp.tsx
@@ -2,8 +2,9 @@
  * useWp hook. Query to retrieve WordPress global stylesheet.
  */
 
-import { gql, useQuery } from '@apollo/client';
+import { gql, useQuery, QueryResult } from '@apollo/client';
 import { omit } from 'lodash';
+import '@assets/css/wordpress.css';
 
 const QUERY_WP_GLOBAL = gql`
 	query WpGlobalQuery {
@@ -11,15 +12,19 @@ const QUERY_WP_GLOBAL = gql`
 	}
 `;
 
+interface WpGlobalQueryResult {
+	wpGlobalStylesheet: string | null;
+}
+
 /**
  * Query to retrieve WordPress global stylesheet.
  *
  * @returns A tuple of the stylesheet string and a query result object.
  */
-const useWp = (): [string | null, any] => {
-	const result = useQuery(QUERY_WP_GLOBAL);
+const useWp = (): [string | null, Omit<QueryResult<WpGlobalQueryResult>, 'data'>] => {
+	const result = useQuery<WpGlobalQueryResult>(QUERY_WP_GLOBAL);
 
-	const { wpGlobalStylesheet } = result.data || {};
+	const { wpGlobalStylesheet = null } = result.data || {};
 
 	return [wpGlobalStylesheet, omit(result, ['data'])];
 };