+++ src/routes/Help.tsx
@@ -1,6 +1,6 @@
 import { Link, Button } from '@chakra-ui/react';
 import { FiFlag } from 'react-icons/fi';
-import ContentView from '@views/ContentView';
+import PageView from '../views/PageView';
 import Page from '@components/Page';
 
 export default function Help() {
@@ -18,7 +18,7 @@ export default function Help() {
 
 	return (
 		<Page title={'Help'} actions={<EmailButton />}>
-			<ContentView postId={'959'} />
+			<PageView postId={'959'} />
 		</Page>
 	);
 }