+++ changelog.md
@@ -9,6 +9,7 @@ with 1.0.0beta.
 - Improve: `prepareUserProfileForGraphQL()` --> `prepareUserProfileForGQL()` for naming convention.
 - Improve: Update Props TS from Index Signature (e.g. `[key: string]: any;`) format to intersection types \(`Props & ListProps`\) where possible.
 - Improve: Minor styling tweaks
+- Improve: Make WP Post retrieval more robust. `ContentView` --> is now `PageView` and `usePostContent` is now `usePage`, which returns a `WPPost` item. This pattern can be reused for other post types!
 
 - TODO: Jobs listing (in progress)
 - TODO: Jobs filtering and search
@@ -16,6 +17,7 @@ with 1.0.0beta.
 - TODO: Job creation
 - TODO: Job edit
 - TODO: Job manage
+- TODO: Refine page content styling ('/help' as example)
 
 ## \[1.2-dev] - Unreleased